third_party/WebKit/LayoutTests/fast/dom/custom/isolated-world.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<body>
<div id="container"><span is="x-a"></span></div>
<script>
description('Tests that accessing custom elements from an isolated world ' +
            'does not cause worlds to collide and destroy the galaxy.');

shouldBeNonNull('window.testRunner', 'this test requires testRunner');

var container = document.querySelector('#container');

// Access an upgrade candidate from an isolated world.
var world = 1;
testRunner.evaluateScriptInIsolatedWorld(
    world,
    'window.container = document.querySelector("#container"); ' +
    'window.a = container.firstElementChild; ' +
    'container.remove();');

// Registering an exception in the isolated world should raise an exception.
result = testRunner.evaluateScriptInIsolatedWorldAndReturnValue(
    world,
    '(function () { ' +
    '  var proto = Object.create(HTMLSpanElement.prototype); ' +
    '  proto.createdCallback = function () { ' +
    '    console.log("entered (isolated)"); ' +
    '  }; ' +
    '  try { ' +
    '    document.register("x-a", {prototype: proto}); ' +
    '    return "register succeeded"; ' +
    '  } catch (e) { ' +
    '    return e.code; ' +
    '  } ' +
    '})()');

shouldBe('result', 'DOMException.NOT_SUPPORTED_ERR',
         'calling register from an isolated world should throw an exception');

// Now access this element from the main world.
a = container.querySelector('[is=x-a]');
shouldBe('a.__proto__', 'HTMLSpanElement.prototype',
         'the main would should not see the usual main world prototype');

// Upgrade from the main world

var proto = Object.create(HTMLSpanElement.prototype);
invocations = [];
proto.enteredDocumentCallback = function () {
    invocations.push('entered (main)');
};
document.register('x-a', {prototype: proto});
shouldBe('invocations', '[]',
         'an element not in a document should not generate entered events');

// Insert from the isolated world

testRunner.evaluateScriptInIsolatedWorld(world++,
                                         'document.body.appendChild(a);');
shouldBe('invocations', '["entered (main)"]',
         'modification in the isolated world should have caused callbacks ' +
         'in the main world');

testPassed('did not crash');
</script>
